ui/pyproject.toml
[project]
name = "elastic-news-ui"
version = "0.1.0"
description = "Web UI for Elastic News newsroom system"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.11.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "mesop>=1.0.0",
    "a2a-sdk>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.2",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
